agent/pyproject.toml
[project]
name = "tensor-langgraph-agent"
version = "0.1.0"
description = "Tensor Telegram bot agent powered by Python + LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=0.4",
  "tavily-python>=0.5",
  "yfinance >= 0.2",
  "fastapi>=0.115",
  "uvicorn>=0.34",
  "python-dotenv>=1",
  "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
